# specs/023-fundacao-competitiva-sazonal/contracts/competitive-foundation.openapi.yaml
openapi: 3.1.0
info:
  title: RinhaDasLendas Competitive Foundation API
  version: 1.0.0
  description: Contrato da feature 023 para Seasons, competições, rodadas, eventos, séries, partidas mínimas e Fearless.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /competicoes:
    get:
      summary: Lista competições com seleção sazonal comum
      parameters:
        - $ref: '#/components/parameters/SeasonIds'
        - $ref: '#/components/parameters/AllSeasons'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Página com seasonsIncluidas, calendarioConfigurado e temporadaAtual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionPage'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /competicoes/{competitionId}:
    parameters:
      - $ref: '#/components/parameters/CompetitionId'
    get:
      summary: Consulta competição por ID
      responses:
        '200':
          description: Competição e rodadas
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Atualiza competição
      x-capability: CanManageCompetitions
      x-resource-condition: Presidente; ou Admin antes de existir Série iniciada no escopo
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompetitionRequest'
      responses:
        '200':
          description: Competição atualizada
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /competicoes/{competitionId}/rodadas:
    parameters:
      - $ref: '#/components/parameters/CompetitionId'
    get:
      summary: Lista rodadas ordenadas
      responses:
        '200':
          description: Rodadas da competição
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Round'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Cria rodada
      x-capability: CanManageCompetitions
      x-resource-condition: Presidente; ou Admin antes de existir Série iniciada no escopo
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoundRequest'
      responses:
        '201':
          description: Rodada criada
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Round'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /competicoes/{competitionId}/ordenacoes-rodadas:
    post:
      summary: Substitui atomicamente a ordem das rodadas
      x-capability: CanManageCompetitions
      x-resource-condition: Presidente; ou Admin antes de existir Série iniciada no escopo
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [rodadaIds]
              properties:
                rodadaIds:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Rodadas reordenadas
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Round'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /competicoes/{competitionId}/regras-publicadas:
    post:
      summary: Publica uma versão imutável de regras
      x-capability: CanManageCompetitions
      x-resource-condition: Exclusivo do Presidente; VicePresidente pendente
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRulesRequest'
      responses:
        '201':
          description: Versão publicada
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RulesVersion'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    ETag:
      description: Versão fraca do recurso, por exemplo W/"7"
      schema:
        type: string
    Location:
      description: URI canônica do recurso criado
      schema:
        type: string
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SeasonIds:
      in: query
      name: temporadaIds
      description: Repetível; mutuamente exclusivo com todas=true
      schema:
        type: array
        uniqueItems: true
        items:
          type: string
          format: uuid
      explode: true
    AllSeasons:
      in: query
      name: todas
      schema:
        type: boolean
        default: false
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 100
    IfMatch:
      in: header
      name: If-Match
      required: true
      description: ETag do recurso; em toda rota de Partida, é a ETag do agregado Série
      schema:
        type: string
...
    CompetitionId:
      in: path
      name: competitionId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Entrada inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Capacidade ausente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso inexistente ou oculto
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Estado, versão ou chave idempotente conflitante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [message, messageCode, errors]
      properties:
        message:
          type: string
        messageCode:
          type: string
        errors:
          type: array
          items:
            type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, messageCode, message]
      properties:
        field:
          type: string
        messageCode:
          type: string
        message:
          type: string
    SeasonEstado:
      type: string
      enum: [Planejada, Ativa, Encerrada]
...
    SeriesFormat:
      type: string
      enum: [Md3, Md5]
    DraftMode:
      type: string
      enum: [Padrao, Fearless]
...
    SeasonSummary:
      type: object
      required: [id, nome, ano, ordemNoAno, dataInicio, dataFimExclusiva, estado, versao]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        ano:
          type: integer
        ordemNoAno:
          type: integer
        dataInicio:
          type: string
          format: date
        dataFimExclusiva:
          type: string
          format: date
        estado:
          $ref: '#/components/schemas/SeasonEstado'
        versao:
          type: integer
          format: int64
...
    UpdateCompetitionRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 120
        codigo:
          type: string
          minLength: 1
          maxLength: 40
        circuitoDiario:
          type: boolean
    CreateRoundRequest:
      type: object
      additionalProperties: false
      required: [nome, ordem]
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 80
        ordem:
          type: integer
          minimum: 1
    PublishRulesRequest:
      type: object
      additionalProperties: false
      required: [formato, modoDraft]
      properties:
        formato:
          $ref: '#/components/schemas/SeriesFormat'
        modoDraft:
          $ref: '#/components/schemas/DraftMode'
...
    CompetitionDetail:
      type: object
      required: [id, seasonId, nome, codigo, circuitoDiario, rodadas, regrasPublicadas, versao, acoesPermitidas]
      properties:
        id:
          type: string
          format: uuid
        seasonId:
          type: string
          format: uuid
        nome:
          type: string
        codigo:
          type: string
        circuitoDiario:
          type: boolean
        rodadas:
          type: array
          items:
            $ref: '#/components/schemas/Round'
        regrasPublicadas:
          type: array
          items:
            $ref: '#/components/schemas/RulesVersion'
        versao:
          type: integer
          format: int64
        acoesPermitidas:
          type: array
          uniqueItems: true
          items:
            type: string
    CompetitionPage:
      type: object
      required: [page, pageSize, items, totalItems, totalPages, calendarioConfigurado, temporadaAtual, seasonsIncluidas]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/CompetitionDetail'
        totalItems:
          type: integer
        totalPages:
          type: integer
        calendarioConfigurado:
          type: boolean
        temporadaAtual:
          oneOf:
            - $ref: '#/components/schemas/SeasonSummary'
            - type: 'null'
        seasonsIncluidas:
          type: array
          items:
            $ref: '#/components/schemas/SeasonSummary'
    Round:
      type: object
      required: [id, competicaoId, nome, ordem, versao]
      properties:
        id:
          type: string
          format: uuid
        competicaoId:
          type: string
          format: uuid
        nome:
          type: string
        ordem:
          type: integer
          minimum: 1
        versao:
          type: integer
          format: int64
    RulesVersion:
      type: object
      required: [id, seasonId, numero, formato, modoDraft, publicadaEm]
      properties:
        id:
          type: string
          format: uuid
        seasonId:
          type: string
          format: uuid
        competicaoId:
          type: [string, 'null']
          format: uuid
        numero:
          type: integer
          minimum: 1
        formato:
          $ref: '#/components/schemas/SeriesFormat'
        modoDraft:
          $ref: '#/components/schemas/DraftMode'
        publicadaEm:
          type: string
          format: date-time
...